+++ code/src/ui_unny.h
@@ -1,0 +1,6 @@
+#ifndef _UI_UNNY_H
+#define _UI_UNNY_H
+
+void setUnnyHUD();
+
+#endif
